requirements.txt
# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# NLP and Transformers
sentence-transformers>=2.2.0
transformers>=4.30.0

# LLM API
groq>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# Model Serialization
joblib>=1.3.0

# Jupyter Notebook (for training)
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: For visualization (if needed)
matplotlib>=3.7.0
seaborn>=0.12.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
